WIP: Convert operation parsing to state machine.
We need a parser that we can pause and resume so that we can split the
parse across multiple chunks.

This is the relatively direct mapping of the imperative parser we had to
one with an explicit state to produce a hopefully more readable diff; a
subsequent commit will regularize the machine into cleaner code.